Feature-complete on ODWY0501: verbs PICK, PUT, and TURN.
Long's "PUT X ON Y" logic is buggy; he creates a dichotomy
between objects that are wearable and those that are
non-wearable, and then treats the command as WEAR or DROP
respectively. And he does this *after* replacing "obj" with
"iobj". So "PUT AXE ON CLOAK" will drop the axe, but
"PUT AXE ON LAMP" will give a snarky response. Again, some
of Long's logic appears to be not just buggy but backwards.
I've fixed the logic.

I've decided that the TURN logic reads more cleanly if I
prohibit TURN LAMP ON LAMP. Long allowed that phrasing, in
addition to TURN LAMP ON and TURN ON LAMP. However, Long
doesn't allow a simple TURN ON or TURN OFF, even though he
helpfully redirects a simple LIGHT to LIGHT LAMP. (Long
also removes Woods' simple one-word ON and OFF commands,
which is slightly annoying.)

The game is now theoretically winnable, although I haven't
verified this myself. I know there is still a bug in the
parser's "Please clarify" logic: entering GET yields "Get
what?", but then answering LAMP doesn't correctly redirect
to GET LAMP. This may be the effect of a misplaced clrlin().