Various small bugfixes.
Investigate the deal with "CONTINUE" as a motion-word and an
action-word, and get it working (although my method of getting
it working is ad-hoc and doesn't emulate Long's logic at all.
Long's logic is still mysterious to me). Ditto "TREE" and "WALL".

Properly initialize the prop values of POLE and TINY_KEY, and
drop their prop values back to 0 in jostle_object().

Properly initialize the flags of lockable and closeable objects.

Long's game starts two of the five dwarves in locations different
from where Woods had them: R_ROTUNDA instead of R_WFISS and
R_SHELL instead of R_COMPLEX.