More bugfixes: Wumpus-related bugs.
The Wumpus puzzle wasn't reachable because I'd forgotten
to add the new exits to R_ELONG. Also, the cloak wasn't
listed as a treasure, so it didn't get its prop value
correctly initialized to 2. Also, the logic for
TAKE CLOAK was missing.

WEARing a carried object should respond with "OK."

I've also deliberately changed the "DROP VASE" output.
In Long's version, it reads:

    >drop pillow, vase
    Velvet pillow
         Dropped.
    The vase is now resting, delicately, on a velvet pillow.
    Ming vase
         Dropped.

In my version, it now reads:

    > drop pillow, vase
    Velvet pillow
        Dropped.
    Ming vase
        The vase is now resting, delicately, on a velvet pillow.

That is, I omit the redundant "Dropped." Ditto for dropping
the bear at the bridge.