Fix Alice/bear-related bugs.
The problem with getting past the dog was that DOG was 102,
which the travel table couldn't handle. I've bumped the modulus
up from 100 to 1000 to fix this.

Added some extra blank lines to the Alice shrink/grow messages,
even though Long's version didn't have them; it just looks a
lot better. Also fixed a grammatical error in a room description.

I'd missed marking a lot of rooms as "forced", which was a
huge error.

The water_at() predicate incorrectly included R_EPIT and
R_WINERY, because I'd blindly copied Long's "liquid at"
predicate and forgot to remove those. This made it impossible
to interact with the wine.

The PHUCE transition was totally broken. It's better now.

Remove trailing whitespace and turn spaces to tabs throughout.

The parser codepath for "THROW BEAR FOOD" doesn't seem to work;
I need to investigate this further.

The BOAT and BEAR don't go through the normal "GET ____" codepath,
but they still contribute to burden(0), so they must be included
in the weight() switch statement. The symptom was an assert(false)
the next time you tried to pick up something.