Fix a few treasure-related bugs found in playtesting.
Killing the dragon assert-failed in shift_words(); this faulty
assertion is now fixed.

The pirate's chest is always ajar (and not hinged).

Ajar objects should have their contents described in the room
description. This notably applies to the chest and the safe.

The maximum number of points per treasure is 5*QK(3),
not 7*QK(3) as I had mistakenly had it.

Points should be awarded for the cask only when it contains
wine (prop value 4); this was a typo on my part.