Fix some mostly endgame-related bugs found in playtesting.
I had an off-by-one in the travel table's size (and the assertion that
was supposed to keep us within the bounds of the travel table). The
symptom was that R_DIDIT's forced move overran the end of the table
and came out as a forced move to 0 (R_LIMBO), which means death.

A period was missing from the Witt's End description.

The repository descriptions hadn't been updated from Woods' version;
Long's version has a phone booth and a rare book mentioned in the
room descriptions. Long's repository also allows the movement verbs
CROSS and ACROSS, and I was missing the travel table entries for
entering the phone booth, which is required in order to win.

The initial "tally" of treasures hadn't been updated from 15 to
Long's 30.

Add more cosmetic blank lines to end-of-game transitions.

Fix a bug in look_around() that caused crashes in the repository.

Fix my buggy translation of Long's BLAST logic.