Fix the next round of bugs found in playtesting.
Long had "echos"; I've fixed it to "echoes".

Long and Woods have "inward"; Knuth changed it to "inwards";
I've changed it back.

The lamp should not give light when it's not "at hand" (namely,
when it's inside a closed sack or chest). This is now fixed.

Long incorrectly printed "You aren't carrying it!" in response
to "WAVE TO FIGURE"; but my code assert-failed instead. This
is now fixed properly. ("Nothing happens.")

My code assert-failed on an input line beginning with "ALL".
It now handles "GET" ("Get what?") "ALL" in the same way that
Long does, but it has arcane differences. TODO: fix these.