Comment on burden() issues; fix an original bug.
Long's original "burden(0)" had issues with recursion, but my
replacement was totally broken due to my mixing up of burden()
and weight(). This new "burden" routine should have much saner
behavior.

Also, comment on some known "features" of Long's PUT logic:
it's possible to PUT X IN Y when X is too heavy to carry
normally, even when Y is currently being carried. I believe
this was an intentional feature of the game logic, to make it
easier for the expert.

The boat should never be "too heavy" to enter. This was an
original bug, now fixed.