Comment on an original bug in FILL CASK [FROM BOTTLE].
Long's FILL logic sets the contents of the vessel based on
IOBJ (the indirect object given by the user or from context),
but it sets the prop value of the vessel based on LOCCON(LOC)
(the liquid available at the current location: oil in the
east pit, wine at the Winery, water everywhere else). This
means that FILL CASK WITH WINE will be accepted when there's
wine present in the bottle, but the resulting CASK object
will be in an inconsistent state: PROP(CASK) will be 0
(indicating water) but the object inside will be WINE+1.

I fixed this bug accidentally when implementing attempt_fill()
from scratch (having decided, rightly, that Long's logic was
too buggy to follow). I didn't even know the bug was present
in Long's version until I tried a speed-run (toting the wine
up the perilous pinnacles in the bottle) and Long's game
refused to give me full points for the cask because its prop
value was still incorrect.

It might be more faithful to the original game (and to mimesis)
for me to prohibit filling the cask from the bottle at all;
but it's such a clever solution to the death-by-pinnacles
problem --- and the alternative solution of bringing the cask
round by sea is so *tedious* --- that I'm inclined to leave
it in. And besides, there's no reason that filling the bottle
from the cask should fail, so the correct logic has to be
there anyway.