Fix(?) the save/restore bug reported by David Watson.
The symptom was that upon restoring, depending on interpreter,
you'd find yourself at either R_ROAD or R_LIMBO (i.e., everything
was restored correctly except for the local variable "loc").
This leads to all kinds of exploits, if "SAVE. RESTORE" can be
used as a teleport-home shortcut!

The problem seems to be that my inline Z-machine-assembly versions
of attempt_save() and attempt_restore() incorrectly overwrote the
local variable "r0" that the compiler's codegen had allocated to
the C variable "loc". The right thing to do is put the save/restore
logic into non-inline functions with their own stack frames and
local variables, so that r0 isn't overwritten by the call to
attempt_restore().

I believe this will fix the issue. It seems to work, in my
tests using Linux "frotz".

Also, I had accidentally provided an extern definition of exit()
in addition to the inline-assembly version. This was redundant
and is now removed.