Disable puts_indented() on the Z-machine.
This is the huge reason that things weren't fitting in RAM.
I'd totally forgotten that puts_indented() interferes with
vbccz's clever optimizing of puts("foo") into @print "foo".
So what I need to do here is implement another compiler hack
before I can re-enable indentation.

Also, stop using atoi() as it's not available on the Z-machine,
and provide an implementation of memset() for the Z-machine.