Fix another bug with "TAKE WATER".
Long redirects "TAKE X" to "REMOVE X FROM Y" whenever X is enclosed
in Y. This is great, except that Long doesn't check for the case
when X is "WATER" and there's water on the ground at the current
location. In Long's original logic, this results in the following
exchange in the starting room:
    > GET WATER
    How?
Not bad; the player can't tell that "How?" really means "I see
there's water in the bottle, but trying to get it out with your
hands is silly."  But with my revised logic, you got this:
    > GET WATER
    You can't get at it.
which is nonsense.

Therefore, I'm redirecting "TAKE X" to "REMOVE X" only when X
is actually here(), which leaves the bottled water out of the
equation unless the bottle is actually in the current room.

As a side effect, this simple change makes "GET WATER" pick up
the water bottle (or cask) if it's in the current room. I
originally thought that the annoying inability to "GET WATER"
was an intentional feature of Long's parser, but perhaps it
was just a bug all along. Anyway, I'm happy enough to leave it
fixed, even though it's a very obvious difference between
my port and MCDO0551.