Add an "-opt-puts=" option to vbccz.
This option will make it possible to get the effect of
    #define puts puts_indented
in Z-code. Of course we could already write exactly that in C,
and the compiler would handle it; but because the function call
wasn't spelled "puts", it wouldn't get the compressed-string-literal
optimization. This blew ODWY0501/ up from 80K to 112K because none
of its strings were getting compressed.

You can supply a Z-machine function name to the -opt-puts= option;
for example "-opt-puts=print_indented". This will cause the C expression
    puts("hello")
to be compiled into
    print_indented("hello^")
on the Z-machine. The default behavior is equivalent to
"-opt-puts=print (string) ".

Printing integers via printf("%d") is still compiled into "print (int)"
regardless of the -opt-puts= option; and embedded Unicode characters
will still be printed via @print_unicode regardless of the -opt-puts=
option. This is fine for my current needs, and unlikely to change.