Make use of the -opt-puts= feature in ODWY0501/.
The most technically interesting part of this patch is in
the new file "indent.inf", function "print_indented". We use
the Z-machine's special stream 3 to print compressed strings
into memory; this gives us access to their individual bytes.
Then we can scan over the bytes looking for newline characters,
the same way we'd do it in C.

The clever "stream 3" approach was recommended to me in the
intfiction forums:
http://www.intfiction.org/forum/viewtopic.php?f=7&t=5989

After this change, adv551.z8 is compilable and playable on
the Z-machine *if* you add the following line after make_loc's
function definition:
    #define make_loc(q,x,l,s,f) make_loc(q,x,s,NULL,f)
This removes all of the long room descriptions from the game,
which shrinks the game enough to fit comfortably in 64K of RAM.

My next task is to refactor the room-description-printing
code so that it doesn't require uncompressed C strings. This
will probably (unfortunately) involve a huge switch statement.