Refactor room descriptions. ODWY0501 is now playable on Parchment!
The huge switch statement isn't so bad, since it's hidden behind
a protective layer of macros. Very little code needed to change.

This approach is of course really stupid for the portable-C
version, but it gets our RAM requirement down to 52873 bytes,
which means adv551.z8 is officially feature-complete and playable!
Now to continue the playtesting...