Rename ODWY0501 to ODWY0551 (the true number of points).
I was originally thinking that I would reconstruct David Long's
original 501-point "Adventure", but I never accumulated enough
primary source material to make a confident reconstruction (and
besides, the Castle of the Elves material is too good to lose).
So, as this is now permanently a 551-point game, the proper
naming convention is really "ODWY0551".