Fix the "EAT AND GET FOOD" family of bugs; fix "Little bird in cage".
The most exploitable of these bugs was "THROW AND GET <some treasure>"
at the troll bridge. However, you could also keep the mushrooms and
cake after eating them (not that that was much use), and recover the
bird after vaporizing or killing it. And there were a few silly
message combinations for nonsense like "KILL AND ANSWER BIRD".

This change preserves Long's logic that rejects certain non-present
nouns immediately at parse time, but also adds a back-up check right
after the "transitive" label to catch objects that were present at
parse time and then disappeared during execution of a player command.

I've taken the liberty of combining the ROD/ROD2 disambiguation code
into "transitive" as well.

Finally, related to multiple objects, I noticed that "GET LAMP AND BIRD"
will print the name of the bird as "Little bird in cage" even when
the bird isn't currently in the cage. This is an original bug, but
I've decided to fix it anyway.