The motion-word BRIDGE (as in "TAKE BRIDGE") should work to move across the fissure in Long's version. There are probably other oversights of this nature in the common parts of the cave; I haven't gone through systematically yet.
Typing "SAY XYZZY" should act just like "XYZZY", without printing
"Okay, 'xyzzy'" first. I had confused the prepping of the "Okay"
message with its actual printing, which happens conditionally
further down in Long's Fortran code. (Unlike Platt's Adventure,
"SAY NORTH" doesn't actually move you north, and so on.)

"Adventure 6" doesn't understand the command "SAY SAINT-MICHEL",
which is annoying because the solution to that puzzle involves
watching as an elf "says the word 'Saint-Michel'". I've decided
that it's best to just add it to the list of magic words that work
when preceded by "SAY".

The parser didn't accept "GO IN" or "GO NORTH", because I'd
forgotten (as usual) that GO is an action-verb, not a motion-verb.
This non-original bug is now fixed.