Fix bugs related to INSERT.
The vending machine's instructions are different in Long's
version from Woods' version. Woods hints at the two-word
command "DROP COINS"; Long insists on "INSERT COINS IN MACHINE".

Due to a bug on my part, "INSERT COINS" would assert-fail.
This is now fixed. The same assert-fail would happen on
"PUT COINS IN" with no indirect object. (In Long's version,
"GET AND PUT COINS IN" enters an infinite loop, for reasons
I haven't investigated.)

"GRIPE" would print two responses due to a "break" that should
have been a "continue". This is now fixed.

"GET IN ALL SACK" was treated as equivalent to "GET SACK",
when it should have been a parse error. This was partly due
to my overloading of the ALL enumerator to also mean ANY,
and partly due to wrong logic in getwds().