Restore Long's original parse for "JUMP STREAM".
The Styx puzzle requires that "JUMP STREAM" and even "JUMP OVER
STREAM" be parsed as synonyms for "JUMP", rather than synonyms
for "STREAM". This must be the reason Long changed Woods' parser,
even though it had the awkward side effect of making "CRAWL NORTH"
a synonym for "CRAWL" ("Which way?") rather than a synonym for
"NORTH" (as in Woods' version). Also notice that "JUMP NORTH"
becomes a synonym for "JUMP", not "NORTH", after this change.