Add a new directory for Luckett and Pike's "Adventure II" (440 points).
This new directory, LUPI0440/, contains Mike Arnautov and Jack Pike's
reconstruction of "Adventure II", exactly as downloaded and un-tarred
from either of these locations:
http://www.mipmip.org/adv440/advIIs.tgz
http://www.ifarchive.org/if-archive/games/source/adv440.tgz

The lineage of this version is:
  WOOD0350 begat LUPI0440, a.k.a. "Adventure II" (Peter Luckett
    and Jack Pike, from 1978 to 1981).
  LUPI0440 was believed lost, but then was reconstructed by
    Mike Arnautov (with Pike's help), circa 2001. The reconstruction
    was done in portable Fortran 77.
  This reconstructed version is called ARNA0440 in Russel Dalenberg's
    taxonomy, and in fact that's what we have here, not the *original*
    Luckett & Pike code (Dalenberg's LUPI0440, which is still lost or
    offline).

A very detailed first-hand account of ARNA0440's history is available
on Mike Arnautov's site:
http://www.mipmip.org/adv440/history.shtml

The new code builds out of the box on Linux as long as you use f77,
but it is insidiously miscompiled by gfortran due to a lack of proper
EXTERNAL directives. (One symptom is that dwarf knives are especially
deadly. Ironically, the other symptom is that if you attempt to enter
an area forbidden to dwarves, the game will crash. This includes most
of Luckett and Pike's new rooms.)