Fix a segfault in LUPI0440.
I don't usually "touch up" the original programs, but in this
case we've got a game-crashing bug that happens completely at
random: if the "little figure" shows up to give one of his
three cryptic hints, but none of them is applicable, then he
tries to RSPEAK an uninitialized message and (on modern
computers) the game segfaults.

The solution is not to RSPEAK anything in that case; move
the last-resort line label "900" down to the next source line.

It's plausible that this could have been a misreproduction
from Pike's paper sources; the original game probably didn't
have this bug in it.