Remove "SLAY" as a synonym for "KILL" in Adv350.
It's unclear what the history is here. "SLAY" does not appear
in my copy of WOOD0350, nor in LUPI0440, but it does appear
independently in WOOD0430 and MCDO0551. It also appears in
KNUT0350 (perhaps imported from WOOD0430, since Woods collaborated
with Knuth on that version) and PLAT0550 (maybe independent,
or maybe cross-pollinated by Mike Arnautov).