Update the MCDO0551 READMEs.
Personal communication from Doug McDonald, 2014-03-04.
A lot of this is still fuzzy, but at least I believe this
version of events to be a lot closer to the truth than
what I'd had in the README before this change.

Mention Robert Hall's Minix port; give more information
about McDonald's and Picton's versions as far as I'm aware.

The significance of the combination to Long's safe (7-22-34)
is that it's the date of John Dillinger's death. Picton
changes it to be different in every game, and also changes
some magic words (such as PHUCE) so that they won't work
until you've "properly" discovered them.