Get adv440 working on the Z-machine.
Namely, by doing the same "locs.h" trick I used for adv551.
In the process, I found a bug: R_CHECK has neither a long
description nor a short description, which causes an assert-fail
in the C version and a bogus "?" printed in the Z-machine version.

Second, fix a bug in adv551: Knuth's adv350 included "forced movement"
locations in the forbidden_to_pirate list, but adv551 didn't (and
neither did adv440, until now). This meant that if the first dwarf
showed up just as you were climbing the beanstalk, you could end up
with the little axe stuck inaccessibly in the "forced movement" location.
This is now fixed.