Fix spelling and grammar errors in ODWY0440.
These are all original to LUPI0440, but I think the game deserves
decent grammar and punctuation. Also, as I did in ODWY0551, normalize
the punctuation on the treasures (namely the ruby) so that all treasures'
descriptions end with an exclamation mark and all non-treasures'
descriptions don't.

The game is now exactly where I want it to be, modulo unknown bugs.