Add WOOD0430 ("generic C source code" version).
This is exactly the contents of
http://www.ifarchive.org/if-archive/games/source/Adventure2.5.tar.gz

If there exists a Fortran version of Woods' "Adventure 2.5", it
is lost in the mists of time. The "copyright" file for this version
claims that it was mechanically translated from Fortran to C, but
not whether this translation happened before or after (or during)
the point(s) where it was expanded from 350 points to 430 points.