Mobile-friendliness: move a newline in the intro message.
Both in Gargoyle and on a mobile browser, there's not enough
room on a line for "Don Knuth,"; it breaks between the words.
Move both words to the next line.

Should SOFT_NL be defined as proposed in 150a267d?
I'm still not sure, as I don't spend enough time playing
in small windows to know if it's a good idea.