Implement assert(x) for Z_MACHINE.
Especially for the versions that didn't get a lot of
testing offline (basically everything after ODWY0350),
it makes sense to have some way to detect assertion
failures in the Parchment version.

No intended change in behavior (unless an assertion
actually does fail); but this does make everything
a little bigger and slower.