Rename {DAIM,POHL}0000->0350, HALL0000->0501
I have now played through Jerry Pohl's "Adventure",
and can confirm that it is 350 points. Russel Dalenberg
taxonomized it as "POHL0000" because of the "split score"
feature, but in fact it's not so much a "split" score as
an "itemized" score. A winning score looks like this:

    Treasures:          218
    Survival:           30
    Getting well in:    25
    Masters section:    25
    Bonus:              45
    Score:              350

Notice that the itemized list adds up to only 343 points;
the "missing" 7 points come from starting the game (2 points),
not quitting (4 points), and the Last Lousy Point (1 point).

Robert Hall's version, interestingly, has 501 points, not
the same 551 points as McDonald's version (with which it is
identical in terms of rooms and items). It shaves 50 points
off of McDonald's score by reducing the number of "fudge"
points that McDonald (and Long, and Woods) give the player
just for showing up. I don't know whether Hall was aware of
David Long's earlier 501-point version; it's possible that
maybe Hall thought he was restoring the maximum score to
what it "should" have been.