Fix a bug in vbccz's squash_stack_frame.
I don't fully understand what goes wrong here, but my stack-frame-squashing
optimization was causing an internal compiler error on the following
trivial program:

    int foo(long p) { return p; }

Some variable in the above program gets allocated with an offset of -2
and a size of 4. I suspect it's the parameter `p` itself, and that
something somewhere in the system is computing the stack offsets of `long`
variables incorrectly, but I haven't figured out where yet.

My Adventure ports don't use the `long` datatype; this came up when
porting Chris Newall's "Chimaera".