New .z8 files, as produced by https://github.com/Quuxplusone/vbccz.
Notice that adv550.z8 hasn't changed in size, and the other three
have gotten bigger. I plan to investigate this, but in the meantime
let's publish these, as I'm interested in getting bug reports about
them (if any bugs are to be found).

This commit does not reflect any source-code change to the games;
it reflects only a change in toolchain.