Update the version of VBCC used to build Adventure.
I'm now tracking my bugfixes and improvements to VBCC
in a separate GitHub repository,
https://github.com/Quuxplusone/vbccz/.
This merely pulls in the changes from that repository
as of commit e5ac730.

As always, Volker Barthelmann's original code is distributed
untouched in the vbcc/ directory (see vbcc/LICENSE), and
my modifications are distributed as a patch file named
"vbcc-patch.diff" in the root directory.