Fix a Knuth bug in scoring reported by Q P Liu!
In https://github.com/Quuxplusone/Advent/pull/1 , Q P Liu
reports that WOOD350 printed end-game texts such as

    YOU SCORED  27 OUT OF A POSSIBLE 350, USING    4 TURNS.
    YOU ARE OBVIOUSLY A RANK AMATEUR.  BETTER LUCK NEXT TIME.
    TO ACHIEVE THE NEXT HIGHER RATING, YOU NEED  3 MORE POINTS.

Notice that 27 + 9 = 36; a score of 35 is *still* an amateur score
in WOOD0350! So KNUT0350 (accidentally?) lowered all the cutoffs
by one point. This has the significant effect of making a 349-point
score "Adventurer Grandmaster" territory in KNUT0350, even though
it was not such in WOOD0350.

Q P Liu has reportedly reported the bug to Knuth.

ODWY0350's end-game text now looks, for example, like this:

    You scored 349 out of a possible 350, using 344 turns.
    Your score puts you in Master Adventurer Class A.
    To achieve the next higher rating, you need 1 more point.

The same bug was present in ODWY0440 (where I have confirmed that
LUPI0440 doesn't have the bug) and ODWY0551 (where I have confirmed
that MCDO0551 doesn't have the bug).

Dave Platt's A-code PLAT0550 redid the table lookup as a nest of "if"s,
putting the cutoffs below the round numbers (e.g. a score of 20 points
is no longer an amateur score).