Fix a `holding_count` bug reported by Jonathan Ellis!
The `OIL` and `WATER` pseudo-objects change their locations from
`R_LIMBO` to `R_INHAND` and back, as you tote and drop the `BOTTLE`.
However, they deliberately do not contribute anything to `holding_count`.
(This is unlike the bird, which does contribute to `holding_count`.)
Therefore, every time we change the location of a liquid, we must be
careful to do it by direct manipulation of `.place`. If we use
`carry`, `drop`, `move`, or `destroy`, we'll change `holding_count`.

The bug here is that if you take the full bottle and then `DROP BOTTLE`,
your holding count incorrectly decreases from 1 to -1. This is my bug;
I introduced it in commit ac411ff4.