Fix an original bug: carrying capacity in the endgame.
This bug is present in both WOOD0350 and KNUT0350.
Similar to the bug fixed in the previous commit, where I was
incorrectly using `move(bottle_contents(), R_LIMBO)` instead
of `obj(bottle_contents()).place = R_LIMBO` and thus decrementing
`holding_count` too many times; this bug is that Woods incorrectly
used `destroy(j)` with `j == WATER`, thus decrementing `holding_count`
too many times.

The extra decrement happens when you're teleported to the endgame.
If you happen to be holding a full bottle when you're teleported,
then your `holding_count` becomes -1, essentially increasing your
carrying capacity from 7 items to 8 items. Coincidentally, there are
exactly 8 totable items in the endgame (the lamp, pillow, cage, bird,
empty bottle, oyster, and both black rods). So you can detect this
bug by attempting to take all 8 items: in a normal game it's impossible
to take an 8th item when you're already toting 7, but if you ended the
main game toting a full bottle, then it will be possible to take
all 8 items in the endgame.

The fix is to send `OIL` and `WATER` manually to limbo first,
exactly as we do when the player dies.