ODWY0440: Destroy dwarf-toted items when the cave is closing.
LUPI0440 does the same thing WOOD0350 does when the sepulchral voice happens:
destroy the dragon, troll, bear, etc., and kill all the dwarves. But in LUPI0440,
not only can dead dwarves be resurrected in R_DWARFQUARTERS, but also, the
dwarf-movement subroutine assumes that any dwarf in R_LIMBO must have been
killed by the player last turn, and so they should drop their toted item
(if any) in the player's location right now. The result is that the horn
(or any other dwarf-toted item) magically teleports into the player's current
location when the sepulchral voice happens:

    A SEPULCHRAL VOICE REVERBERATING THROUGH THE CAVE, SAYS, "CAVE CLOSING
    SOON.  ALL ADVENTURERS EXIT IMMEDIATELY THROUGH MAIN OFFICE."

    THERE IS A LITTLE HORN HERE.

    YOU'RE IN THE DIRTY PASSAGE.

    THERE IS A LITTLE HORN HERE.

This behavior is pretty inexplicable from the player's point of view,
and isn't indispensable to gameplay.

Also, one could use this trick to get the vase into the sewage maze,
which is supposed to be impossible (a dropped vase will shatter instead
of disappearing in the sewage). In fact, one could use this to get
*any* toted item to appear atop the sewage, which isn't really supposed
to happen.

For all these reasons, let's eliminate this behavior in ODWY0440.