MCDO0551: Compile with -fno-automatic
This is required in order to make the kaleidoscope maze work:
the local variable `KALFLG` needs to retain its value between
calls to the `TRAVL` subroutine. Another way to do this would
have been to add `SAVE KALFLG` at the top of the subroutine;
but I'm not sure if there might be other variables also expecting
to be saved, elsewhere in the code. So `-fno-automatic` is safer
and more historically accurate.

The Makefile for PLAT0550 also uses `gfortran`, but I'm not aware
that the lack of `-fno-automatic` is causing any problems for PLAT0550,
so I'm leaving it alone.

It's unclear to me whether `f77` needs a similar command-line option.
I have no way to test `f77`, so I'm leaving it alone.