Update the version of VBCC used to build Adventure
This pulls in the changes from https://github.com/Quuxplusone/vbccz/
as of commit c06a3b7 (unauthorized-version-9h),
which is based on vbcc 0.9hP2 (March 2022).

As always, Volker Barthelmann's original code is distributed
untouched in the vbcc/ directory (see vbcc/LICENSE), and
my modifications are distributed as a patch file named
"vbcc-patch.diff" in the root directory.