ANON1400: Original PDFs and transcription.
advent.pdf and adven2.pdf come via email from Rick Hammerstone, who played
this game "on some variation of an HP minicomputer in either summer 1981
or 1982" at "a summer computer camp at the local high school". He says
that the programming language is "HP Time-Shared BASIC".

In the transcriptions "advent.ocr.txt" and "adven2.ocr.txt",
blank lines separate pages of the original document.
There are no blank lines in the original at all.

The transcriptions preserve the wonky "indentation" of whole lines in the
original, even though we confidently guess that this was somehow a glitch
in the 1980s printer hardware, and doesn't reflect anything about the
actual original files. It's trivial to remove those leading spaces (and
the blank lines, and the "ADVENT"/"ADVEN2" header lines) if you want to
get as close as possible to the original files.

The consistent double-space before "NOT" and "END" probably indicates
that the original files were stored in a tokenized format.