Commit sources for ANON0501 and OSKA0501
These are simply downloaded from the IF Archive, unzipped,
and checked into git. I haven't tried to build them or anything.

ANON0501: https://www.ifarchive.org/if-archive/games/source/advent.tar.Z
OSKA0501: https://www.ifarchive.org/if-archive/games/source/advent-5.2.2-5.tar.gz

The fairly strong conjecture here is that both of these are very small
modifications on top of the lost LONG0501, and the common intersection
of ANON0501 and OSKA0501 is basically LONG0501 in a nutshell.
(This corresponds to "MCDO0551 minus the Castle of the Elves puzzle.")